requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.39.0
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==5.0.1
dnspython==2.6.1
gunicorn==21.2.0
idna==3.4
oauthlib==3.2.2
packaging==23.2
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.2
python-dotenv==1.0.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
